+++ tests/test_create_observation/log/test_compile_log_type.py
