+++ tests/test_create_observation/log/test_get_log_list_from_url.py
@@ -6,7 +6,7 @@ from typing import Optional
 from freezegun import freeze_time
 import pytest
 from fetch_data.api_data.log_data.fetch_log_data import get_log_list_from_url
-from tests.test_create_observation.test_create_log_observation.assert_log_record import (
+from tests.test_create_observation.log.assert_log_record import (
     assert_log_record,
 )
