+++ tests/test_create_observation/software_version/data/Mediawiki_UniversalLanguageSelector.html
