+++ tests/test_create_observation/software_version/test_constants.py
@@ -1,0 +1,3 @@
+"""Constants"""
+
+DATA_DIRECTORY = "tests/test_create_observation/software_version/data"
