--- tests/test_create_observation/test_create_software_version_observation/test_constants.py
@@ -1,5 +1,0 @@
-"""Constants"""
-
-DATA_DIRECTORY = (
-    "tests/test_create_observation/test_create_software_version_observation/data"
-)
